pokeapi-mongo-s3-airflow-etl/requirements.txt
pymongo==3.11
apache-airflow-providers-mongo==3.0.0
aiohttp==3.9.5